fix: refactored route docs og
